move back logic into GameplayMenuOverlay
